refactor: Calculator 클래스 코드 분리
